requirements.txt
# Core dependencies
crewai
crewai-tools
boto3
PyYAML
python-dotenv
langchain_aws